New debug and server features along with small tweaks to the UI
Added new window to display client info when a user is clicked in the Server ControlPanel.
Added ability to kick users.
Implemented enhancement described in this [issue](#16).
Changed MenuBar to be more streamlined.
Fixed issue on Client close that would cause an exception to be thrown and the application to hang.